core_function/Unused/requirements.txt
numpy<2.0.0
ffmpeg-python==0.2.0
openai-whisper==20231117
pyannote.audio==3.1.1
pydub==0.25.1
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
transformers<4.36.0
huggingface-hub>=0.16.4
scikit-learn>=1.0.0